% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{as.enrich_res}
\alias{as.enrich_res}
\title{as enrich_res object}
\usage{
as.enrich_res(gsea_res)
}
\arguments{
\item{gsea_res}{gsea_res from KO_gsea}
}
\value{
enrich_res object
}
\description{
as enrich_res object
}
