% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_net.R
\name{calculate_NS}
\alias{calculate_NS}
\title{Calculate the NS (Node Score) for KEGG pathway networks}
\usage{
calculate_NS(path_index, lambda = 0.5)
}
\arguments{
\item{path_index}{data.frame from `pathway_net_index`}

\item{lambda}{Numeric, a parameter for weighting the normalized degree and downstream node count. Default is `0.5`.}
}
\value{
A data frame with additional columns for normalized degree, normalized downstream node count, and NS score.
}
\description{
Calculate the NS (Node Score) for KEGG pathway networks
}
