% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_net.R
\name{calculate_ZN_score}
\alias{calculate_ZN_score}
\title{Calculate ZN Scores for Pathways}
\usage{
calculate_ZN_score(all_NS_res, cor_df)
}
\arguments{
\item{all_NS_res}{A data frame containing pathway information with columns:
Pathway_id, name, and NS.}

\item{cor_df}{A data frame containing correlation data with columns: name and cor.}
}
\value{
A data frame with columns: Pathway_id, ZN_score, p_value, p_value2.
}
\description{
This function calculates ZN scores, p-values, and secondary p-values for each pathway
based on correlation data and NS values.
}
\seealso{
calculate_NS
}
