% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{export_report_table}
\alias{export_report_table}
\title{Export report score result tables}
\usage{
export_report_table(reporter_res, dir_name, overwrite = FALSE)
}
\arguments{
\item{reporter_res}{a reporter_score object or rs_by_cm object}

\item{dir_name}{the directory to save the report tables}
}
\value{
No return value
}
\description{
Export report score result tables
}
