% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{extract_cluster}
\alias{extract_cluster}
\title{Extract one cluster from rs_by_cm object}
\usage{
extract_cluster(rsa_cm_res, cluster = 1)
}
\arguments{
\item{rsa_cm_res}{rs_by_cm object}

\item{cluster}{integer}
}
\value{
reporter_score object
}
\description{
Extract one cluster from rs_by_cm object
}
