% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{get_features}
\alias{get_features}
\alias{get_KOs}
\title{get features in a modulelist}
\usage{
get_features(map_id = "map00010", ko_stat = NULL, modulelist = NULL)
}
\arguments{
\item{map_id}{map_id in modulelist}

\item{ko_stat}{NULL or ko_stat result from \code{\link{pvalue2zs}}}

\item{modulelist}{NULL or customized modulelist dataframe, must contain 'id','K_num','KOs','Description' columns. Take the `KOlist` as example, use \code{\link{custom_modulelist}}.}
}
\value{
KOids, or data.frame with these KOids.
}
\description{
get features in a modulelist
}
\examples{
get_features(map_id = "map00010")
}
\seealso{
Other modulelist: 
\code{\link{custom_modulelist}()},
\code{\link{custom_modulelist_from_org}()}
}
\concept{modulelist}
