% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{get_org_pathway}
\alias{get_org_pathway}
\title{Get pathway information for an organism from kegg}
\usage{
get_org_pathway(
  org = "hsa",
  RDS_file = NULL,
  download = TRUE,
  download_dir = NULL
)
}
\arguments{
\item{org}{kegg organism, listed in https://www.genome.jp/kegg/catalog/org_list.html, default, "hsa"}

\item{RDS_file}{path of a org.RDS file if you saved before.}

\item{download}{save the .keg file?}

\item{download_dir}{where to save the .keg file?}
}
\description{
Get pathway information for an organism from kegg
}
