% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{get_reporter_score}
\alias{get_reporter_score}
\title{Calculate reporter score}
\usage{
get_reporter_score(
  ko_stat,
  type = c("pathway", "module")[1],
  threads = 1,
  KOlist_file = NULL,
  modulelist = NULL,
  perm = 1000,
  verbose = TRUE
)
}
\arguments{
\item{ko_stat}{ko_stat result from \code{\link{pvalue2zs}}}

\item{type}{"pathway" or "module" for default KOlist_file.}

\item{threads}{threads}

\item{KOlist_file}{default NULL, use the internal file. Or you can upload your .rda file from \code{\link{make_KO_list}}}

\item{modulelist}{NULL or customized modulelist dataframe, must contain "id","K_num","KOs","Description" columns. Take the `KOlist` as example, use \code{\link{custom_modulelist}}.}

\item{perm}{permutation number, default: 1000}

\item{verbose}{logical}
}
\value{
reporter_res dataframe
}
\description{
Calculate reporter score
}
\examples{
\donttest{
data(KO_abundance_test)
ko_pvalue=ko.test(KO_abundance,"Group",metadata)
ko_stat=pvalue2zs(ko_pvalue,mode="directed")
reporter_s1=get_reporter_score(ko_stat)
}
}
