% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsa_kegg_pathway}
\alias{hsa_kegg_pathway}
\title{pathway information for "hsa"}
\description{
pathway information for "hsa"
}
\seealso{
Other data: 
\code{\link{CPDlist}},
\code{\link{Compound_htable}},
\code{\link{GOlist}},
\code{\link{KO_htable}},
\code{\link{KOlist}},
\code{\link{Module_htable}},
\code{\link{Pathway_htable}},
\code{\link{mmu_kegg_pathway}}
}
\concept{data}
