% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ko.test}
\alias{ko.test}
\title{Wilcox-test or t.test for KO-abundance table}
\usage{
ko.test(
  kodf,
  group,
  metadata = NULL,
  threads = 1,
  p.adjust.method = "BH",
  verbose = TRUE
)
}
\arguments{
\item{kodf}{KO_abundance table, rowname is ko id (e.g. K00001), colnames is samples}

\item{group}{The compare group (two category) in your data, one column name of metadata when metadata exist or a vector whose length equal to columns number of kodf.}

\item{metadata}{sample information dataframe contains group}

\item{threads}{default 1}

\item{p.adjust.method}{p.adjust.method, see \code{\link[stats]{p.adjust}}}

\item{verbose}{logical}
}
\value{
ko_pvalue dataframe
}
\description{
Wilcox-test or t.test for KO-abundance table
}
\examples{
\donttest{
data("KO_abundance_test")
ko_pvalue=ko.test(KO_abundance,"Group",metadata)
}
}
