% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KO_abundance}
\alias{KO_abundance}
\alias{metadata}
\title{The KOs abundance table and group table.}
\description{
The KOs abundance table and group table.

The KOs abundance table and group table.
}
\seealso{
Other test_data: 
\code{\link{genedf}},
\code{\link{reporter_score_res}}
}
\concept{test_data}
