% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{KO_gsa_internal}
\alias{KO_gsa_internal}
\title{gene set analysis}
\usage{
KO_gsa_internal(
  kodf,
  group,
  metadata = NULL,
  resp.type = "Two class unpaired",
  modulelist = NULL,
  p.adjust.method = "BH",
  verbose = TRUE,
  perm = 1000
)
}
\description{
gene set analysis
}
\examples{
debug(KO_gsa_internal)
KO_gsa_internal(KO_abundance,"Group",metadata)
}
