% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{ko_test}
\alias{ko_test}
\title{Wilcox-test for KO-abundance table,}
\usage{
ko_test(
  kodf,
  group,
  metadata = NULL,
  vs_group = NULL,
  verbose = T,
  threads = 1,
  p.adjust.method = "BH"
)
}
\arguments{
\item{kodf}{KO_abundance table, rowname is ko id (e.g. K00001),colnames is samples}

\item{group}{one column name of metadata when metadata exist or a vector length equal to columns number of kodf}

\item{metadata}{sample information dataframe contains group}

\item{vs_group}{vs_group should contains group levels}

\item{verbose}{logical}

\item{threads}{default 1}
}
\value{
ko_pvalue dataframe
}
\description{
Wilcox-test for KO-abundance table,
}
\examples{
data(KO_test)
ko_pvalue=ko_test(KO_abundance,"Group",Group_tab)
}
