% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_net.R
\name{load_all_pathway_NS}
\alias{load_all_pathway_NS}
\title{Load all KEGG pathway NS results}
\usage{
load_all_pathway_NS(org = NULL, verbose = TRUE)
}
\arguments{
\item{org}{Character, the KEGG organism code (e.g., "hsa" for human). If `NULL`, uses "ko" as the default prefix for pathway IDs.}

\item{verbose}{Logical, whether to print messages about the loading process. Default is `TRUE`.}
}
\value{
A data frame containing the NS scores for all pathways, with columns: Pathway_id, name, NS.
}
\description{
Load all KEGG pathway NS results
}
\seealso{
calculate_NS
}
