% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_CARDinfo}
\alias{load_CARDinfo}
\title{Load the CARDinfo (from CARD)}
\usage{
load_CARDinfo(envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
CARDinfo in `.GlobalEnv`
}
\description{
Load the CARDinfo (from CARD)
}
