% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_CAZy_info}
\alias{load_CAZy_info}
\title{Load the CAZy_info (from CAZy database)}
\usage{
load_CAZy_info(verbose = TRUE)
}
\arguments{
\item{verbose}{logical}
}
\value{
CAZy_info
}
\description{
Load the CAZy_info (from CAZy database)
}
