% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_Compound_htable}
\alias{load_Compound_htable}
\title{Load the Compound_htable (from KEGG)}
\usage{
load_Compound_htable(envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
Compound_htable in `.GlobalEnv`
}
\description{
Load the Compound_htable (from KEGG)
}
