% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_CPDlist}
\alias{load_CPDlist}
\title{Load the CPDlist (from KEGG)}
\usage{
load_CPDlist(CPDlist_file = NULL, envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
CPDlist in `.GlobalEnv`
}
\description{
Load the CPDlist (from KEGG)
}
