% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_Enzyme_info}
\alias{load_Enzyme_info}
\title{Load the Enzyme_info (from ExPASy Enzyme database)}
\usage{
load_Enzyme_info(verbose = TRUE)
}
\arguments{
\item{verbose}{logical}
}
\value{
Enzyme_info
}
\description{
Load the Enzyme_info (from ExPASy Enzyme database)
}
