% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_GOlist}
\alias{load_GOlist}
\alias{load_GOinfo}
\title{Load the GOlist (from 'GO' database)}
\usage{
load_GOlist(verbose = TRUE)

load_GOinfo(verbose = TRUE)
}
\arguments{
\item{verbose}{logical}
}
\value{
GOlist

GOinfo
}
\description{
Load the GOlist (from 'GO' database)

Load the GOinfo (from GO)
}
\examples{
load_GOlist()
}
