% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_KO_htable}
\alias{load_KO_htable}
\title{Load the KO_htable (from KEGG)}
\usage{
load_KO_htable(envir = .GlobalEnv)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}
}
\value{
KO_htable in `.GlobalEnv`
}
\description{
Load the KO_htable (from KEGG)
}
