% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_KOlist}
\alias{load_KOlist}
\title{Load the KOlist (from KEGG)}
\usage{
load_KOlist(KOlist_file = NULL, envir = .GlobalEnv)
}
\arguments{
\item{KOlist_file}{NULL or your `KOlist.rda` results from `make_KO_list`}

\item{envir}{`.GlobalEnv` (default) or `environment()`}
}
\value{
KOlist in `.GlobalEnv`
}
\description{
Load the KOlist (from KEGG)
}
