% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_Module_htable}
\alias{load_Module_htable}
\title{Load the Module_htable (from KEGG)}
\usage{
load_Module_htable(envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
Module_htable in `.GlobalEnv`
}
\description{
Load the Module_htable (from KEGG)
}
