% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_org_pathway}
\alias{load_org_pathway}
\title{Load the pathway information for an organism (from KEGG)}
\usage{
load_org_pathway(org = "hsa", envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{org}{kegg organism, listed in https://www.genome.jp/kegg/catalog/org_list.html, default, "hsa"}

\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
KOlist in `.GlobalEnv`
}
\description{
Load the pathway information for an organism (from KEGG)
}
