% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_net.R
\name{load_pathway_xml_ls}
\alias{load_pathway_xml_ls}
\title{Load KEGG pathway XML file list}
\usage{
load_pathway_xml_ls(org = NULL, verbose = TRUE)
}
\arguments{
\item{org}{kegg organism, listed in https://www.genome.jp/kegg/catalog/org_list.html such as "hsa", default NULL means ko.}

\item{verbose}{Logical, whether to print messages about the loading process. Default is `TRUE`.}
}
\value{
A list of KEGG pathway XML files, where each element is a `tbl_graph` or `igraph` object.
}
\description{
Load KEGG pathway XML file list
}
