% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{make_KO_list}
\alias{make_KO_list}
\title{Prepare KOlist from KEGG}
\usage{
make_KO_list(download_dir, output = NULL)
}
\arguments{
\item{download_dir}{directory of your download file}

\item{output}{default, download_dir/KOlist.rda}
}
\description{
Prepare KOlist from KEGG
}
