% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{modify_description}
\alias{modify_description}
\title{Modify the pathway description before plotting}
\usage{
modify_description(
  reporter_res,
  pattern = " - Homo sapiens (human)",
  replacement = ""
)
}
\arguments{
\item{reporter_res}{reporter_res}

\item{pattern}{str, like " - Homo sapiens (human)"}

\item{replacement}{str, like ""}
}
\value{
reporter_res
}
\description{
Modify the pathway description before plotting
}
\examples{
data("reporter_score_res")
modify_description(reporter_score_res, pattern = " - Homo sapiens (human)")
}
