% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{parse_enzyme_dat}
\alias{parse_enzyme_dat}
\title{Parse enzyme.dat file into a structured data frame}
\usage{
parse_enzyme_dat(file_path)
}
\arguments{
\item{file_path}{Path to the enzyme.dat file.}
}
\value{
A data frame with columns: EC, Name, Aliases, Reaction, Comments, SwissProt_Refs.
}
\description{
Parse enzyme.dat file into a structured data frame
}
