% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pathway_htable}
\alias{Pathway_htable}
\title{Pathway htable from 'KEGG'}
\description{
Pathway htable from 'KEGG'
}
\seealso{
Other data: 
\code{\link{CPDlist}},
\code{\link{Compound_htable}},
\code{\link{GOlist}},
\code{\link{KO_htable}},
\code{\link{KOlist}},
\code{\link{Module_htable}},
\code{\link{hsa_kegg_pathway}},
\code{\link{mmu_kegg_pathway}}
}
\concept{data}
