% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_net.R
\name{pathway_net_index}
\alias{pathway_net_index}
\title{Create a pathway network index data.frame}
\usage{
pathway_net_index(path_net_c)
}
\arguments{
\item{path_net_c}{A `tbl_graph` or `igraph` object representing a KEGG pathway network.}
}
\value{
A data frame containing vertex indices and their attributes, including in-degree and out-degree.
}
\description{
Create a pathway network index data.frame
}
\seealso{
plot_pathway_net
}
