% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cm_res}
\alias{plot.cm_res}
\title{Plot c_means result}
\usage{
\method{plot}{cm_res}(
  x,
  filter_membership,
  mode = 1,
  show.clust.cent = TRUE,
  show_num = TRUE,
  ...
)
}
\arguments{
\item{x}{a cm_res object}

\item{filter_membership}{filter membership}

\item{mode}{1~2}

\item{show.clust.cent}{show cluster center?}

\item{show_num}{show number of each cluster?}

\item{...}{additional}
}
\value{
ggplot
}
\description{
Plot c_means result
}
