% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_c_means}
\alias{plot_c_means}
\title{Plot c_means result}
\usage{
plot_c_means(
  rsa_cm_res,
  filter_membership,
  mode = 1,
  show.clust.cent = TRUE,
  show_num = TRUE,
  ...
)
}
\arguments{
\item{rsa_cm_res}{a cm_res object}

\item{filter_membership}{filter membership 0~1.}

\item{mode}{1~2}

\item{show.clust.cent}{show cluster center?}

\item{show_num}{show number of each cluster?}

\item{...}{additional}
}
\description{
Plot c_means result
}
