% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_enrich_res}
\alias{plot_enrich_res}
\title{Plot enrich_res}
\usage{
plot_enrich_res(
  enrich_res,
  mode = 1,
  str_width = 50,
  padj_threshold = 0.05,
  facet_level = FALSE,
  facet_anno = NULL,
  facet_str_width = 15,
  ...
)
}
\arguments{
\item{enrich_res}{enrich_res object}

\item{mode}{plot style: 1~2}

\item{str_width}{default: 50}

\item{padj_threshold}{p.adjust threshold}

\item{facet_level}{facet plot if the type is "pathway" or "module"}

\item{facet_anno}{annotation table for facet, two columns, first is level summary, second is pathway id.}

\item{facet_str_width}{str width for facet label}

\item{...}{add}
}
\value{
ggplot
}
\description{
Plot enrich_res
}
