% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_features_distribution}
\alias{plot_features_distribution}
\alias{plot_KOs_distribution}
\title{plot the Z-score of features distribution}
\usage{
plot_features_distribution(
  reporter_res,
  map_id,
  text_size = 4,
  text_position = NULL,
  rug_length = 0.04
)
}
\arguments{
\item{reporter_res}{result of `reporter_score`}

\item{map_id}{the pathway or module id}

\item{text_size}{text_size=4}

\item{text_position}{text_position, e.g. c(x=3,y=0.4)}

\item{rug_length}{rug_length=0.04}
}
\value{
ggplot
}
\description{
plot the Z-score of features distribution
}
\examples{
data("reporter_score_res")
plot_features_distribution(reporter_score_res, map_id = c("map05230", "map03010"))
}
