% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_htable}
\alias{plot_htable}
\title{plot htable levels}
\usage{
plot_htable(type = "ko", select = NULL, htable = NULL)
}
\arguments{
\item{select_ko}{select kos}
}
\value{
ggplot
}
\description{
plot htable levels
}
\examples{
data("KO_abundance_test")
plot_htable(select=rownames(KO_abundance))
}
