% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_KO_htable}
\alias{plot_KO_htable}
\title{plot KO_htable levels}
\usage{
plot_KO_htable(select_ko = NULL)
}
\arguments{
\item{select_ko}{select kos}
}
\value{
ggplot
}
\description{
plot KO_htable levels
}
\examples{
data("KO_abundance_test")
plot_KO_htable(rownames(KO_abundance))
}
