% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{plot_KOs_box}
\alias{plot_KOs_box}
\title{plot KOs boxplot}
\usage{
plot_KOs_box(
  KO_abundance,
  Group,
  Group_tab,
  map_id = "map00780",
  select_ko = NULL,
  ...
)
}
\description{
plot KOs boxplot
}
