% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_report}
\alias{plot_report}
\title{Plot the reporter_res}
\usage{
plot_report(
  reporter_res,
  rs_threshold = 1.64,
  mode = 1,
  y_text_size = 13,
  str_width = 50,
  Pathway_description = TRUE
)
}
\arguments{
\item{reporter_res}{result of `get_reporter_score`}

\item{rs_threshold}{plot threshold vector, default:1.64}

\item{mode}{1～2 plot style.}

\item{y_text_size}{y_text_size}

\item{str_width}{str_width to wrap}

\item{Pathway_description}{show KO description rather than KO id.}
}
\value{
ggplot
}
\description{
Plot the reporter_res
}
\examples{
data("reporter_score_res")
plot_report(reporter_score_res,rs_threshold=c(2,-2),y_text_size=10,str_width=40)
}
