% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_significance}
\alias{plot_significance}
\title{Plot the significance of pathway}
\usage{
plot_significance(reporter_res, map_id)
}
\arguments{
\item{reporter_res}{result of `get_reporter_score` or `reporter_score`}

\item{map_id}{the pathway or module id}
}
\value{
ggplot
}
\description{
Plot the significance of pathway
}
\examples{
data("reporter_score_res")
plot_significance(reporter_score_res,map_id=c("map05230","map03010"))
}
