% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.reporter_score}
\alias{print.reporter_score}
\title{Print reporter_score}
\usage{
\method{print}{reporter_score}(x, ...)
}
\arguments{
\item{x}{reporter_score}

\item{...}{add}
}
\value{
No value
}
\description{
Print reporter_score
}
