% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{print.rs_by_cm}
\alias{print.rs_by_cm}
\title{Print rs_by_cm}
\usage{
\method{print}{rs_by_cm}(x, ...)
}
\arguments{
\item{x}{rs_by_cm}

\item{...}{add}
}
\value{
No value
}
\description{
Print rs_by_cm
}
