% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{reporter_score}
\alias{reporter_score}
\title{One step to get the reporter score of your KO abundance table.}
\usage{
reporter_score(
  kodf,
  group,
  metadata = NULL,
  mode = c("mixed", "directed")[1],
  verbose = TRUE,
  threads = 1,
  p.adjust.method = "BH",
  type = c("pathway", "module")[1],
  perm = 1000,
  KOlist_file = NULL,
  modulelist = NULL
)
}
\arguments{
\item{kodf}{KO_abundance table, rowname is ko id (e.g. K00001),colnames is samples.}

\item{group}{The compare group (two category) in your data, one column name of metadata when metadata exist or a vector whose length equal to columns number of kodf. And you can use factor levels to change order.}

\item{metadata}{sample information dataframe contains group.}

\item{mode}{"mixed" or "directed", see details in \code{\link{pvalue2zs}}.}

\item{verbose}{logical}

\item{threads}{default 1}

\item{p.adjust.method}{p.adjust.method, see \code{\link[stats]{p.adjust}}}

\item{type}{"pathway" or "module"}

\item{perm}{permutation number, default: 1000.}

\item{KOlist_file}{default NULL, use the internal file. Or you can upload your .rda file from \code{\link{make_KO_list}}.}

\item{modulelist}{NULL or customized modulelist dataframe, must contain "id","K_num","KOs","Description" columns. Take the `KOlist` as example, use \code{\link{custom_modulelist}}.}
}
\value{
reporter_score object：
\item{kodf}{your input KO_abundance table}
\item{ko_pvalue}{ko statistics result contains p.value}
\item{ko_stat}{ko statistics result contains p.value and z_score}
\item{reporter_s}{the reporter score in each pathway}
\item{modulelist}{default KOlist or customized modulelist dataframe}
\item{group}{The compare group (two category) in your data}
\item{metadata}{sample information dataframe contains group}
}
\description{
One step to get the reporter score of your KO abundance table.
}
\examples{
\donttest{
reporter_score_res=reporter_score(KO_abundance,"Group",metadata,mode="directed")
}
}
