% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporterScore-package.R
\docType{package}
\name{ReporterScore-package}
\alias{ReporterScore}
\alias{ReporterScore-package}
\title{ReporterScore: Generalized Reporter Score-Based Enrichment Analysis for Omics Data}
\description{
Inspired by the classic 'RSA', we developed the improved 'Generalized Reporter Score-based Analysis (GRSA)' method, implemented in the R package 'ReporterScore', along with comprehensive visualization methods and pathway databases. 'GRSA' is a threshold-free method that works well with all types of biomedical features, such as genes, chemical compounds, and microbial species. Importantly, the 'GRSA' supports multi-group and longitudinal experimental designs, because of the included multi-group-compatible statistical methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Asa12138/ReporterScore}
  \item Report bugs at \url{https://github.com/Asa12138/ReporterScore/issues}
}

}
\author{
\strong{Maintainer}: Chen Peng \email{bfzede@gmail.com} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})

}
\keyword{internal}
