% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{RSA_by_cm}
\alias{RSA_by_cm}
\title{Reporter score analysis after C-means clustering}
\usage{
RSA_by_cm(
  kodf,
  group,
  metadata = NULL,
  k_num = NULL,
  filter_var = 0.7,
  seed = 1234,
  verbose = TRUE,
  method = "pearson",
  ...
)
}
\arguments{
\item{kodf}{KO_abundance table, rowname is ko id (e.g. K00001),colnames is samples.}

\item{group}{The comparison groups (at least two categories) in your data, one column name of metadata when metadata exist or a vector whose length equal to columns number of kodf. And you can use factor levels to change order.}

\item{metadata}{sample information data.frame contains group}

\item{k_num}{if NULL, perform the \code{\link[pctax]{cm_test_k}}, else an integer}

\item{filter_var}{see \code{\link[pctax]{c_means}}}

\item{seed}{set.seed}

\item{verbose}{verbose}

\item{method}{method from \code{\link{reporter_score}}}

\item{...}{additional arguments for \code{\link{reporter_score}}}
}
\value{
rs_by_cm
}
\description{
Reporter score analysis after C-means clustering
}
\examples{
\donttest{
rsa_cm_res=RSA_by_cm(KO_abundance,"Group2",metadata,method = "pearson")
}
}
