% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{transform_modulelist}
\alias{transform_modulelist}
\title{Transform a modulelist to a list}
\usage{
transform_modulelist(mymodulelist)
}
\arguments{
\item{mymodulelist}{mymodulelist}
}
\description{
Transform a modulelist to a list
}
