% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_net.R
\name{update_all_pathway_NS}
\alias{update_all_pathway_NS}
\title{Update all KEGG pathway networks and calculate their NS scores}
\usage{
update_all_pathway_NS(pathway_xml_ls = NULL, org = NULL, lambda = 0.5)
}
\arguments{
\item{pathway_xml_ls}{A list of KEGG pathway XML files, where each element is a `tbl_graph` or `igraph` object. If `NULL`, it will load the existing XML files.}

\item{org}{Character, the KEGG organism code (e.g., "hsa" for human). If `NULL`, uses "ko" as the default prefix for pathway IDs.}

\item{lambda}{Numeric, a parameter for weighting the normalized degree and downstream node count. Default is `0.5`.}
}
\description{
Update all KEGG pathway networks and calculate their NS scores
}
\seealso{
calculate_NS
}
