% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_CAZy_info}
\alias{update_CAZy_info}
\title{Update CAZy_info from CAZy database}
\usage{
update_CAZy_info(
  download_dir = NULL,
  fam_activities_file = NULL,
  fam_subfam_ec_file = NULL
)
}
\arguments{
\item{download_dir}{Directory to download the CAZy data files. If NULL, a temporary directory will be used.}

\item{fam_activities_file}{the CAZy family activities file.}

\item{fam_subfam_ec_file}{the CAZy family-subfamily-EC file.}
}
\value{
No value
}
\description{
Update CAZy_info from CAZy database
}
