% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_Enzyme_info}
\alias{update_Enzyme_info}
\title{Update Enzyme_info from ExPASy Enzyme database}
\usage{
update_Enzyme_info(
  download_dir = NULL,
  enzyme_dat_file = NULL,
  enzclass_file = NULL
)
}
\arguments{
\item{download_dir}{Directory to download the Enzyme data files. If NULL, a temporary directory will be used.}

\item{enzyme_dat_file}{Path to the manually downloaded `enzyme.dat` file. If NULL, the file will be downloaded.}

\item{enzclass_file}{Path to the manually downloaded `enzclass.txt` file. If NULL, the file will be downloaded.}
}
\value{
No value
}
\description{
Update Enzyme_info from ExPASy Enzyme database
}
