% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_GOinfo}
\alias{update_GOinfo}
\title{update_GOinfo}
\usage{
update_GOinfo(download_dir = NULL, obo_file = NULL)
}
\arguments{
\item{download_dir}{download_dir}

\item{obo_file}{obo_file from http://current.geneontology.org/ontology/go.obo}
}
\description{
update_GOinfo
}
