% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_htable}
\alias{update_htable}
\title{update some htable from KEGG}
\usage{
update_htable(type, keg_file = NULL, download = FALSE, download_dir = NULL)
}
\arguments{
\item{type}{"ko", "module", "pathway", "compound" ...}

\item{keg_file}{path of a .keg file, such as ko00001.keg from https://www.genome.jp/kegg-bin/download_htext?htext=ko00001&format=htext.}

\item{download}{save the .keg file?}

\item{download_dir}{where to save the .keg file?}
}
\description{
update some htable from KEGG
}
