% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_KO_file}
\alias{update_KO_file}
\title{Update the KO2Pathway and KO2Module files}
\usage{
update_KO_file(pack_dir = NULL)
}
\arguments{
\item{pack_dir}{the ReporterScore package location, detect automatically.}
}
\description{
Download links:

\link{https://rest.kegg.jp/link/pathway/ko}
\link{https://rest.kegg.jp/list/pathway}
\link{https://rest.kegg.jp/link/module/ko}
\link{https://rest.kegg.jp/list/module}
\link{https://rest.kegg.jp/list/ko}

you can also download yourself the use \code{\link{make_KO_list}} to get a KOlist object
}
\examples{
update_KO_file()

}
