% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_KO_htable}
\alias{update_KO_htable}
\title{update_KO_htable from KEGG}
\usage{
update_KO_htable(file = NULL, download_dir = NULL)
}
\arguments{
\item{file}{ko00001.keg from https://www.genome.jp/kegg-bin/download_htext?htext=ko00001&format=htext}

\item{download_dir}{download_dir}
}
\value{
KO_htable
}
\description{
update_KO_htable from KEGG
}
