% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_Module_htable}
\alias{update_Module_htable}
\title{update Module_htable from KEGG}
\usage{
update_Module_htable(file = NULL, download_dir = NULL)
}
\arguments{
\item{file}{ko00002.keg from https://www.genome.jp/kegg-bin/download_htext?htext=ko00002&format=htext}

\item{download_dir}{download_dir}
}
\value{
Module_htable
}
\description{
update Module_htable from KEGG
}
