% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{custom_modulelist}
\alias{custom_modulelist}
\title{Build a custom moduleist}
\usage{
custom_modulelist(pathway2ko, pathway2desc = NULL)
}
\arguments{
\item{pathway2ko}{user input annotation of Pathway to KO mapping, a data.frame of 2 column with pathway and ko.}

\item{pathway2desc}{user input of Pathway TO Description mapping, a data.frame of 2 column with pathway and description.}
}
\value{
a custom modulelist
}
\description{
Build a custom moduleist
}
\examples{
mydat=data.frame(pathway=paste0("PATHWAY",rep(1:2,each=5)),ko=paste0("K",1:10))
mymodulelist=custom_modulelist(mydat)
mymodulelist
}
