% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{custom_modulelist_from_org}
\alias{custom_modulelist_from_org}
\title{Custom modulelist from a specific organism}
\usage{
custom_modulelist_from_org(org = "hsa", feature = "ko", verbose = TRUE)
}
\arguments{
\item{org}{kegg organism, listed in https://www.genome.jp/kegg/catalog/org_list.html, default, "hsa"}

\item{feature}{one of "ko", "gene", "compound"}

\item{verbose}{logical}
}
\value{
modulelist
}
\description{
Custom modulelist from a specific organism
}
\examples{
custom_modulelist_from_org(org = "hsa", feature = "ko")
}
