% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{get_reporter_score}
\alias{get_reporter_score}
\title{Calculate reporter score}
\usage{
get_reporter_score(
  ko_stat,
  mode = c("pathway", "module")[1],
  verbose = T,
  threads = 1,
  KOlist_file = NULL,
  perm = 1000
)
}
\arguments{
\item{ko_stat}{ko_stat result from \code{\link{pvalue2zs}}}

\item{mode}{"pathway" or "module"}

\item{verbose}{logical}

\item{threads}{threads}

\item{KOlist_file}{default NULL, use the internal file. Or you can upload your .rda file from \code{\link{make_KO_list}}}
}
\value{
reporter_res dataframe
}
\description{
Calculate reporter score
}
\examples{
data(KO_test)
ko_pvalue=ko_test(KO_abundance,"Group",Group_tab)
ko_stat=pvalue2zs(ko_pvalue,mode="directed")
reporter_s=get_reporter_score(ko_stat)
}
