% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{KO_enrich}
\alias{KO_enrich}
\title{Perform KO enrichment analysis}
\usage{
KO_enrich(ko_pvalue, p.adjust.method = "BH")
}
\arguments{
\item{ko_pvalue}{ko_pvalue dataframe from \code{\link{ko_test}}.}

\item{p.adjust.method}{The method used for p-value adjustment (default: "BH").}
}
\value{
A data frame containing the enrichment results.
}
\description{
This function performs KO enrichment analysis using the clusterProfiler package.
}
\examples{
a=KO_enrich(ko_pvalue)
plot(a)

}
