% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{KO_gsea}
\alias{KO_gsea}
\title{Perform KO gene set enrichment analysis}
\usage{
KO_gsea(ko_pvalue, p.adjust.method = "BH")
}
\arguments{
\item{ko_pvalue}{ko_pvalue dataframe from \code{\link{ko_test}}.}

\item{p.adjust.method}{The method used for p-value adjustment (default: "BH").}
}
\description{
This function performs KO enrichment analysis using the clusterProfiler package.
}
\examples{
a=KO_gsea(ko_pvalue)
gseaplot(e,geneSetID = e@result$ID[1])
}
