% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{KO_safe}
\alias{KO_safe}
\title{Perform Significance Analysis of Function and Expression}
\usage{
KO_safe(
  reporter_res,
  verbose = TRUE,
  perm = 1000,
  C.matrix = NULL,
  p.adjust.method = "BH",
  ...
)
}
\arguments{
\item{reporter_res}{The input reporter result.}

\item{verbose}{If TRUE, print verbose messages. Default is TRUE.}

\item{perm}{The number of permutations. Default is 1000.}

\item{C.matrix}{The contrast matrix. Default is NULL, and it will be generated from the module list.}

\item{p.adjust.method}{Method for p-value adjustment. Default is "BH".}

\item{...}{Additional parameters to be passed to \code{\link[safe]{safe}} function.}
}
\value{
A data frame containing SAFE results for KO enrichment.
}
\description{
Perform Significance Analysis of Function and Expression
}
\examples{
\donttest{
## use `safe` from the `safe` package.
if (requireNamespace("safe")) {
  data("reporter_score_res")
  safe_res <- KO_safe(reporter_score_res,
    verbose = TRUE,
    perm = 200, p.adjust.method = "none"
  )
}
}
}
