% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KOlist}
\alias{KOlist}
\title{The KOlist used for enrichment.}
\format{
four columns in each data.frame.
\describe{
\item{id}{"map0010" or "M00001"}
\item{K_num}{contians how many KOs in this pathway or module}
\item{KOs}{KOs name}
\item{Description}{the description of this pathway or module}
}
}
\description{
an list contains two data.frame named pathway and module.
}
\seealso{
Other data: 
\code{\link{CPDlist}},
\code{\link{Compound_htable}},
\code{\link{GOlist}},
\code{\link{KO_htable}},
\code{\link{Module_htable}},
\code{\link{Pathway_htable}},
\code{\link{hsa_kegg_pathway}},
\code{\link{mmu_kegg_pathway}}
}
\concept{data}
