% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_CARDinfo}
\alias{load_CARDinfo}
\title{Load the CARDinfo (from CARD database)}
\usage{
load_CARDinfo(verbose = TRUE)
}
\arguments{
\item{verbose}{logical}
}
\value{
CARDinfo
}
\description{
Load the CARDinfo (from CARD database)
}
\examples{
load_CARDinfo()
}
