% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_GOlist}
\alias{load_GOlist}
\title{Load the GOlist (from GO)}
\usage{
load_GOlist(envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
KO_htable in `.GlobalEnv`
}
\description{
Load the GOlist (from GO)
}
