% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_GOlist}
\alias{load_GOlist}
\alias{load_GOinfo}
\title{Load the GOlist (from GO database)}
\usage{
load_GOlist(envir = .GlobalEnv, verbose = TRUE)

load_GOinfo(envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
GOlist in `.GlobalEnv`

GOinfo in `.GlobalEnv`
}
\description{
Load the GOlist (from GO database)

Load the GOinfo (from GO)
}
\examples{
load_GOlist()
}
