% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_htable}
\alias{load_htable}
\title{Load the KO_htable (from KEGG)}
\usage{
load_htable(type, envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{type}{"ko", "module", "pathway", "compound" ...}

\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
KO_htable in `.GlobalEnv`
}
\description{
Load the KO_htable (from KEGG)
}
