% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{load_KO_desc}
\alias{load_KO_desc}
\title{Load the KO description (from KEGG)}
\usage{
load_KO_desc(envir = .GlobalEnv, verbose = TRUE)
}
\arguments{
\item{envir}{`.GlobalEnv` (default) or `environment()`}

\item{verbose}{logical}
}
\value{
KO description in `.GlobalEnv`
}
\description{
Load the KO description (from KEGG)
}
