% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{make_KO_list}
\alias{make_KO_list}
\title{Prepare KOlist}
\usage{
make_KO_list(dir, output = NULL)
}
\arguments{
\item{dir}{directory of your download file}

\item{output}{default, dir/KOlist.rda}
}
\description{
Prepare KOlist
}
