% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_enrich_res}
\alias{plot_enrich_res}
\alias{plot.enrich_res}
\title{Plot enrich_res}
\usage{
plot_enrich_res(
  enrich_res,
  mode = 1,
  padj_threshold = 0.05,
  show_ID = FALSE,
  Pathway_description = TRUE,
  facet_level = FALSE,
  facet_anno = NULL,
  str_width = 50,
  facet_str_width = 15,
  ...
)

\method{plot}{enrich_res}(
  x,
  mode = 1,
  padj_threshold = 0.05,
  show_ID = FALSE,
  Pathway_description = TRUE,
  facet_level = FALSE,
  facet_anno = NULL,
  str_width = 50,
  facet_str_width = 15,
  ...
)
}
\arguments{
\item{enrich_res}{enrich_res object}

\item{mode}{plot style: 1~2}

\item{padj_threshold}{p.adjust threshold}

\item{show_ID}{show pathway id}

\item{Pathway_description}{show KO description rather than KO id.}

\item{facet_level}{facet plot if the type is "pathway" or "module"}

\item{facet_anno}{annotation table for facet, two columns, first is level summary, second is pathway id.}

\item{str_width}{default: 50}

\item{facet_str_width}{str width for facet label}

\item{...}{add}

\item{x}{enrich_res object}
}
\value{
ggplot

ggplot
}
\description{
Plot enrich_res

Plot enrich_res
}
\examples{
data("reporter_score_res")
enrich_res <- KO_enrich(reporter_score_res)
plot(enrich_res)
}
