% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_htable}
\alias{plot_htable}
\title{Plot htable levels}
\usage{
plot_htable(type = "ko", select = NULL, htable = NULL)
}
\arguments{
\item{type}{"ko", "module", "pathway", "compound"}

\item{select}{select ids}

\item{htable}{custom a htable}
}
\value{
ggplot
}
\description{
Plot htable levels
}
\examples{
\donttest{
data("KO_abundance_test")
plot_htable(select = rownames(KO_abundance))
}
}
