% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{plot_KOs_in_pathway}
\alias{plot_KOs_in_pathway}
\title{Plot KOs trend in one pathway or module}
\usage{
plot_KOs_in_pathway(
  map_id = "map00780",
  ko_stat = ko_stat,
  box_color = get_cols(2),
  line_color = c("skyblue3", "red2", "grey")
)
}
\arguments{
\item{map_id}{the pathway or module id}

\item{ko_stat}{ko_stat result from \code{\link{pvalue2zs}}}

\item{box_color}{box and point color}

\item{line_color}{line color}
}
\value{
ggplot
}
\description{
Plot KOs trend in one pathway or module
}
\examples{
plot_KOs_in_pathway(map_id="map00780",ko_stat = ko_stat)
}
