% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{plot_report}
\alias{plot_report}
\title{Plot the reporter_res}
\usage{
plot_report(
  reporter_res,
  rs_threshold = 1.64,
  mode = 1,
  y_text_size = 13,
  str_width = 50
)
}
\arguments{
\item{reporter_res}{result of `get_reporter_score`}

\item{rs_threshold}{plot threshold vector, default:1.64}

\item{y_text_size}{y_text_size}

\item{str_width}{str_width to wrap}

\item{color}{color vector c('#e31a1c','#47B0D9')}
}
\value{
ggplot
}
\description{
Plot the reporter_res
}
\examples{
data(KO_test)
ko_pvalue=ko_test(KO_abundance,"Group",Group_tab)
ko_stat=pvalue2zs(ko_pvalue,mode="directed")
reporter_s=get_reporter_score(ko_stat)
plot_report(reporter_s,rs_threshold=c(2,-7),y_text_size=10,str_width=40)
}
