% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporterScore-package.R
\docType{package}
\name{ReporterScore-package}
\alias{ReporterScore}
\alias{ReporterScore-package}
\title{ReporterScore: Generalized Reporter Score-based Enrichment Analysis for Diverse Omics Data}
\description{
The original reporter-score method comes from Patil, K. R. et al. PNAS 2005. In this package, we provide two reporter-score methods (mixed/directed) to get reporter-score for KEGG pathway or modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Asa12138/ReporterScore}
  \item Report bugs at \url{https://github.com/Asa12138/ReporterScore/issues}
}

}
\author{
\strong{Maintainer}: Peng Chen \email{pengchen2001@zju.edu.cn}

}
\keyword{internal}
