% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_CARDinfo}
\alias{update_CARDinfo}
\title{update_CARDinfo}
\usage{
update_CARDinfo(download_dir = NULL, card_data = NULL)
}
\arguments{
\item{download_dir}{download_dir}

\item{card_data}{card_data from https://card.mcmaster.ca/download/0/broadstreet-v3.2.8.tar.bz2}
}
\description{
update_CARDinfo
}
