% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{update_KO_file}
\alias{update_KO_file}
\title{Update the KO2Pathway and KO2Module files from KEGG}
\usage{
update_KO_file(download_dir = NULL)
}
\arguments{
\item{download_dir}{the ReporterScore user dir location, detect automatically by `tools::R_user_dir("ReporterScore")`.}
}
\description{
Download links:

\code{https://rest.kegg.jp/link/pathway/ko}
\code{https://rest.kegg.jp/list/pathway}
\code{https://rest.kegg.jp/link/module/ko}
\code{https://rest.kegg.jp/list/module}

you can also download yourself the use \code{\link{make_KO_list}} to get a KOlist object
}
