% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_net.R
\name{update_pathway_xml_ls}
\alias{update_pathway_xml_ls}
\title{Download KEGG pathway XML files and create networks}
\usage{
update_pathway_xml_ls(download_dir)
}
\arguments{
\item{download_dir}{Directory to save the downloaded XML files.}
}
\value{
No value
}
\description{
Download KEGG pathway XML files and create networks
}
